* process.c (run_exec_dup2): fixed type.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@25031 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ process.c
@@ -1925,7 +1925,7 @@ run_exec_dup2(VALUE ary, VALUE save, char *errmsg, size_t errmsg_buflen)
         int oldfd;
         int newfd;
         long older_index;
-        int num_newer;
+        long num_newer;
     } *pairs = 0;
 
     n = RARRAY_LEN(ary);
